requirements.txt
setuptools>=49.2.1
python-magic>=0.4.22
python-dateutil>=2.8.1
pathlib>=1.0.1
argparse>=1.4.0